Patched merged shaders

+++ shaders/chunk.vert
@@ -15,7 +15,7 @@ uniform mat4 m_view;
 uniform mat4 m_model;
 
 out vec2 uv;
-out int out_face_id;
+out float out_face_id;
 out float out_block_id;
 
 const vec2 uv_coords[8] = vec2[8](
@@ -32,7 +32,7 @@ const int uv_indices[18] = int[18](
 void main() {
     int uv_index = int(gl_VertexID % 6 + face_id * 6);
     uv = uv_coords[uv_indices[uv_index]];
-    out_face_id = int(face_id);
+    out_face_id = face_id;
     out_block_id = block_id;
 
     gl_Position = m_proj * m_view * m_model * vec4(in_position, 1.0);
